Fix image link crawling
Fixes #1

Maximum concurrency level is only 16 now because there is single
`t1.daumcdn.net` server and aria2 doesn't allow to use more. Though it
seems to be much faster CDN than `cfile.uf.tistory.com` (resolves to
`t1.int.daumcdn.net`) as per June 2018 so 16 threads should be enough.

Also the old one might be deprecated in future, so it seems to be the
best option.